% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{edi_update_package}
\alias{edi_update_package}
\title{Update EDI repository package}
\usage{
edi_update_package(eml_path, edi_cred, edi_env = "staging")
}
\arguments{
\item{eml_path}{Path to an EML document for the EDI data package.
Should include scope, packagid, and revision number (e.g.
knb-lter-jrn.999.3)}

\item{edi_cred}{EDI credentials in a list}

\item{edi_env}{EDI repository environment (staging, production, or
development)}
}
\description{
This function takes a path to an EML file (with an EDI package identifier)
and updates the package in the given EDI environment (default is 'staging').
The given EML document must be present in the working directory. NOTE that
if this isn't working one might need to upload EML and data at
portal-s.edirepository.org
}
