% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_from_mb}
\alias{eml_from_mb}
\title{EML metadata from metabase}
\usage{
eml_from_mb(pkgid, dbname, dbcred)
}
\arguments{
\item{pkgid}{Number or numeric vector of dataset IDs to query}

\item{dbname}{Name of the database ('metabase') to query metadata from}

\item{dbcred}{Credentials for the database (a list)}
}
\value{
An EML-formatted list of metadata
}
\description{
This function queries a metabase and formats metadata to the EML spec
}
