% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{ents_to_s3}
\alias{ents_to_s3}
\title{Put entities in an S3 bucket}
\usage{
ents_to_s3(ents, bucket_name, multi_part = FALSE)
}
\arguments{
\item{ents}{A list of dataTable or otherEntity filenames}

\item{bucketname}{Name of the s3 bucket to upload to}
}
\description{
This function takes a list of data entities and uploads them to an s3
bucket
}
