% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetAttributes_df}
\alias{format_DataSetAttributes_df}
\title{Create the DataSetAttributes and related tables from a dataframe}
\usage{
format_DataSetAttributes_df(dsid, df)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{df}{A dataframe object or list of dataframes}
}
\value{
A named list containing four dataframes formatted to match the
lter_metabase.'DataSetAttributes', lter_metabase.'ListCodes',
lter_metabase.'DataSetAttributeEnumeration', and
lter_metabase.'DataSetAttributeMissingCodes' tables.
}
\description{
Create the DataSetAttributes and related tables from a dataframe
}
