% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetPersonnel}
\alias{format_DataSetPersonnel}
\title{Create the DataSetPersonnel and related tables from an EML list (emld object)}
\usage{
format_DataSetPersonnel(dsid, eml)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}
}
\value{
A named list containing dataframes formatted to match the
lter_metabase."DataSetPersonnel", lter_metabase."ListPeople", and
lter_metabase."ListPeopleID" tables
}
\description{
Create the DataSetPersonnel and related tables from an EML list (emld object)
}
