% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetTemporal}
\alias{format_DataSetTemporal}
\title{Create the DataSetTemporal table from an EML list (emld object)}
\usage{
format_DataSetTemporal(dsid, eml)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}
}
\value{
A named list containing a dataframe formatted to match the
lter_metabase."DataSetTemporal" table
}
\description{
Create the DataSetTemporal table from an EML list (emld object)
}
