% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{get_entities}
\alias{get_entities}
\title{Get dataTable and otherEntity filenames}
\usage{
get_entities(eml.list)
}
\arguments{
\item{eml.list}{An EML list object}
}
\value{
A list of filenames
}
\description{
This function extracts the filenames for dataTables and otherEntities
}
