% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{increment_edi_revision}
\alias{increment_edi_revision}
\title{Update EML revision numbers using EDI repository}
\usage{
increment_edi_revision(eml_list, edi_env = "staging")
}
\arguments{
\item{eml_list}{R list of EML-schema-formatted metadata}

\item{edi_env}{Name of EDI repository environment ('staging', 'production',
or 'development')}
}
\value{
An EML-schema-formatted list of metadata with revised package
revision numbers.
}
\description{
This function queries the EDI repository to retrieve the current
revision number for a data package there, and then updates an EML
list accordingly.
}
