% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{ingress_table}
\alias{ingress_table}
\title{Append a formatted dataframe to an LTER metabase table}
\usage{
ingress_table(conn, tablename, df, schema = "lter_metabase")
}
\arguments{
\item{conn}{Connection to an LTER Metabase instance}

\item{tablename}{Name (char) of a table in the given schema in LTER Metabase}

\item{df}{A dataframe formatted to match the given LTER Metabase table}

\item{schema}{Name (char) of the LTER Metabase schema to target
("lter_metabase" by default)}
}
\value{
Nothing
}
\description{
Append a formatted dataframe to an LTER metabase table
}
