% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{insert_datasource_xml}
\alias{insert_datasource_xml}
\title{Insert dataSource XML elements into an EML document}
\usage{
insert_datasource_xml(eml_list, xml_file, islist = TRUE)
}
\arguments{
\item{eml_list}{An EML-schema-formatted R list (EML R package compliant)}

\item{xml_file}{XML file with one or more \if{html}{\out{<dataSource>}} elements (full path)}
}
\value{
An updated EML-schema-formatted R list (EML R package compliant)
}
\description{
This function inserts \if{html}{\out{<dataSource>}} elements into a methodstep for an
EML document. Supply an XML file with a list of data sources and it
will be inserted into the first methodstep.
}
