% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{load_destination_cred}
\alias{load_destination_cred}
\title{Get dataset destination credentials}
\usage{
load_destination_cred(path)
}
\arguments{
\item{path}{Path to the 'destination_keys.R' file containing API keys}
}
\value{
edicred - a list containing EDI credentials
}
\description{
This function reads credentials for any data repositories and s3 buckets
in a jerald destination credentials file.
}
