% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{load_edi_cred}
\alias{load_edi_cred}
\title{Get EDI credentials}
\usage{
load_edi_cred(path)
}
\arguments{
\item{path}{Path to the 'edi_keys.R' file containing your EDI credentials
credentials}
}
\value{
edicred - a list containing metabase credentials
}
\description{
This function reads credentials for an EDI user into the edicred
list variable.
}
