% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{load_s3_cred}
\alias{load_s3_cred}
\title{Get s3 bucket API keys}
\usage{
load_s3_cred(path)
}
\arguments{
\item{path}{Path to the 's3_keys.R' file containing s3 API keys}
}
\value{
1 if successful
}
\description{
This function reads credentials for an s3 bucket into environmental
variable for access by the aws.s3 package.
}
