% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{metabase_connect}
\alias{metabase_connect}
\title{Connect to an LTER Metabase}
\usage{
metabase_connect(mbcred_path)
}
\arguments{
\item{mbcred_path}{Path to a file containing credentials for an LTER Metabase instance}
}
\value{
A connection to the LTER Metabase
}
\description{
Connect to an LTER Metabase
}
