% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{migrate_eal_dir}
\alias{migrate_eal_dir}
\title{Migrate an EAL dataset directory to a jerald directory}
\usage{
migrate_eal_dir(eal_dir, jerald_dir)
}
\arguments{
\item{eal_dir}{Source dataset directory, in EAL format, to migrate from}

\item{jerald_dir}{Destination dataset directory, in jerald format, to
migrate to}
}
\description{
WARNING - this is fairly jornada-specific use at your own risk
}
\details{
This will move the files from an EMLassemblyline-formatted dataset directory
to a jerald-formatted dataset directory (as described in
\verb{jerald/inst/template/}). The jerald-formatted destination should be
created first using \code{jerald::template_dataset_dir()} using a name distinct
from the source directory.

All files from the EAL source directory will be copied into a new
\verb{EAL_archive/} directory in the jerald destination directory. Then, this
function copies any "build" R scripts and data entities into the top level
of the jerald directory, moves old metadata files (\code{.dsd}, etc) into
\verb{metadata_docs/}, and moves EML files to \verb{eml/}.

Optionally you may remove the source dataset directory at the end but
YOU SHOULD VERY CAREFULLY CHECK THE OUTPUT BEFORE ANSWERING YES!
}
