% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{parse_edi_pid}
\alias{parse_edi_pid}
\title{Get the EDI dataset revision number from an EML document}
\usage{
parse_edi_pid(eml_list, parse = "scope")
}
\arguments{
\item{eml_list}{R list of EML-schema-formatted metadata}

\item{parse}{A string designating which part of the EDI package ID to return
('scope', 'dataset', or 'revision')}
}
\value{
The requested portion of the EDI package ID
}
\description{
Any EML document created for the EDI repository should have a packageId
field. Parse out the parts of that.
}
