% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{remove_ents_s3}
\alias{remove_ents_s3}
\title{Remove entities from S3 bucket}
\usage{
remove_ents_s3(ents, bucket_name)
}
\arguments{
\item{ents}{A list of dataTable or otherEntity filenames in an s3 bucket}

\item{bucketname}{Name of the s3 bucket to modify}
}
\description{
This function takes a list of data entities and deletes them from an s3
bucket
}
