% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{return_if_node_exists}
\alias{return_if_node_exists}
\title{Get an EML element at a specified address if it exists in the EML list (emld)}
\usage{
return_if_node_exists(qlist, nodeid = 1)
}
\arguments{
\item{qlist}{An EML list address to query}

\item{nodeid}{Name (char) or index (int) of the element to check existence of
and return if present in qlist}
}
\value{
NA if address is NULL or nodeid doesn't exist, otherwise the element
value at qlist/nodeid
}
\description{
Get an EML element at a specified address if it exists in the EML list (emld)
}
