% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{return_list_if_single}
\alias{return_list_if_single}
\title{Return single EML elements at a specified address as a list}
\usage{
return_list_if_single(qlist)
}
\arguments{
\item{qlist}{An EML list address to query}

\item{nodeid}{Name (char) or index (int) of the element to check existence of
and return if present in qlist}
}
\value{
NULL if qlist is NULL (doesn't exist), qlist if it is an unnamed list
(list of element types), and list(qlist) if it contains named elements
(single, unlisted element type).
}
\description{
This helps standardize EML element types into lists for iteration
}
