% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{suggest_Keywords}
\alias{suggest_Keywords}
\title{Suggest keywords from metabase ListKeywords to match ListKeywords from eml}
\usage{
suggest_Keywords(from_eml, from_metabase)
}
\arguments{
\item{from_eml}{A dataframe formatted to match the lter_metabase."ListKeywords"
table - taken from an EML file}

\item{from_metabase}{A dataframe formatted to match the
lter_metabase."ListKeywords" table - taken from an LTER metabase table}
}
\value{
A table merging eml- and metabase-derived ListKeywords tables on "Keyword"
}
\description{
Suggest keywords from metabase ListKeywords to match ListKeywords from eml
}
