% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{update_dataset_edi}
\alias{update_dataset_edi}
\title{Update a dataset at the EDI repository}
\usage{
update_dataset_edi(
  datasetid,
  mb_name,
  mb_cred,
  edi_cred,
  edi_env = "staging",
  publish = FALSE,
  bucket_name = Sys.getenv("AWS_S3_BUCKETNAME")
)
}
\arguments{
\item{datasetid}{ID number of the dataset to find in metabase and
update in EDI}

\item{mb_name}{name of the metabase database in the postgres cluster}

\item{mb_cred}{list of credentials for the metabase postgres cluster}

\item{edi_cred}{list of credentials to use for EDI}

\item{edi_env}{name of the EDI environment to update (staging, production,
or development)}

\item{publish}{boolean value - publish if TRUE, end before s3 upload if
FALSE}

\item{bucket_name}{name of the s3 bucket to push data entities to}
}
\description{
This function updates an existing dataset (or "data package") at the EDI
research data repository using metadata derived from an LTER Metabase.
The user must supply credentials for the metabase and EDI (see
\code{load_metabase_cred} and \code{load_destination_cred} functions), and
appropriate database and EDI environment names.
}
\details{
The basic process is to
\enumerate{
\item Pull metadata for the dataset from the metabase using \code{MetaEgress}
\item Query for the current dataset revision in the EDI environment
\item Write an EML document for the next revision to go to EDI
\item Push EML data entities from the working directory to an s3 bucket
\item Push the EML document to EDI, which triggers PASTA to pull data from
the s3 bucket and update the data package.
}
}
