% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{update_edi_package}
\alias{update_edi_package}
\title{Update EDI repository package}
\usage{
update_edi_package(emlpkgid, edicred, edienv = "staging")
}
\arguments{
\item{emlpkgid}{Package identifier for EML document and EDI data package.
Should include scope, packagid, and revision number (e.g.
knb-lter-jrn.210000000.3)}

\item{edicred}{EDI credentials in a list}

\item{edienv}{EDI repository environment (staging, production, or
development)}
}
\description{
This function takes an EDI package identifier and updates the package on EDI.
The correspondingly named EML document must be present in the working
directory. NOTE that if this isn't working one might have to upload EML
(but not data) at portal.edirepository.org
}
