% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edi.R
\name{edi_update_package}
\alias{edi_update_package}
\title{Update EDI repository package}
\usage{
edi_update_package(packageid, edi.cred, edi.env = "staging")
}
\arguments{
\item{edi.cred}{EDI credentials in a list}

\item{edi.env}{EDI repository environment (staging, production, or
development)}

\item{emlpkgid}{Package identifier for EML document and EDI data package.
Should include scope, packagid, and revision number (e.g.
knb-lter-jrn.210000000.3)}
}
\description{
This function takes an EDI package identifier and updates the package on EDI.
The correspondingly named EML document must be present in the working
directory. NOTE that if this isn't working one might have to upload EML
and data at portal-s.edirepository.org
}
