% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml_from_mb}
\alias{eml_from_mb}
\title{Create EML list from an LTER metabase}
\usage{
eml_from_mb(datasetid, mb.name, mb.cred)
}
\arguments{
\item{datasetid}{Number or numeric vector of dataset IDs to query}

\item{mb.name}{Name of the LTER metabase to query metadata from}

\item{mb.cred}{Credentials for mb.name (a list, see \code{load_metabase_cred})}
}
\value{
An EML-schema-formatted list of metadata
}
\description{
This function queries an LTER metabase and formats the metadata into an
R list in the EML schema
}
