% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSet}
\alias{format_DataSet}
\title{Create the DataSet table from an EML list (emld object)}
\usage{
format_DataSet(
  dsid,
  eml,
  revnum = 0,
  outpath = getwd(),
  boilerplate_id = "jrn_default"
)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}

\item{outpath}{The path to output an abstract markdown file to. Defaults to
current working directory}

\item{boilerplate_id}{The identifier of a "boilerplate" record in metabase,
which will set the project tree and other elements.}
}
\value{
A named list containing a dataframe formatted to match the
lter_metabase.'DataSet' table
}
\description{
Create the DataSet table from an EML list (emld object)
}
