% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetEntities}
\alias{format_DataSetEntities}
\title{Create the DataSetEntities table from an EML list (emld object)}
\usage{
format_DataSetEntities(dsid, eml)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}
}
\value{
A named list containing a dataframe formatted to match the
lter_metabase.'DataSetEntities' table
}
\description{
NOTE: currently this does not handle otherEntities
}
