% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetKeywords}
\alias{format_DataSetKeywords}
\title{Create the DataSetKeywords and related tables from an EML list (emld object)}
\usage{
format_DataSetKeywords(dsid, eml)
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}
}
\value{
A named list containing dataframes formatted to match the
lter_metabase.'DataSetKeywords' and lter_metabase.'ListKeywords' tables
}
\description{
Create the DataSetKeywords and related tables from an EML list (emld object)
}
