% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ingress.R
\name{format_DataSetMethod}
\alias{format_DataSetMethod}
\title{Create the DataSetMethod table from an EML list (emld object)}
\usage{
format_DataSetMethod(dsid, eml, outpath = getwd())
}
\arguments{
\item{dsid}{The dataset id value, which is the primary key for lter_metabase}

\item{eml}{An emld object derived from an EML file}

\item{outpath}{The path to output a methods markdown file to. Defaults to
current working directory}
}
\value{
A named list containing a dataframe formatted to match the
lter_metabase.'DataSetMethod' table
}
\description{
Create the DataSetMethod table from an EML list (emld object)
}
