% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{get_eml_entities}
\alias{get_eml_entities}
\title{Get dataTable and otherEntity filenames from EML}
\usage{
get_eml_entities(eml.list)
}
\arguments{
\item{eml.list}{An EML-schema-formatted R list}
}
\value{
A list of data entity filenames found in the EML
}
\description{
This function extracts the data entity filenames (dataTables and
otherEntities) from an EML-schema-formatted list and returns a vector
of the filenames (useful for upload to s3 buckets).
}
