% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{insert_methodstep_md}
\alias{insert_methodstep_md}
\title{Insert markdown methodstep element in an EML document}
\usage{
insert_methodstep_md(eml.list, md_file, islist = TRUE)
}
\arguments{
\item{eml.list}{An EML-schema-formatted R list (EML R package compliant)}

\item{md_file}{A markdown methods file (full path)}
}
\value{
An EML-schema-formatted R list (EML R package compliant)
}
\description{
This function extracts the data entity filenames (dataTables and
otherEntities) from an EML-schema-formatted list and returns a vector
of the filenames (useful for upload to s3 buckets).
}
