% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{load_metabase_cred}
\alias{load_metabase_cred}
\title{Get LTER Metabase credentials}
\usage{
load_metabase_cred(pathname)
}
\arguments{
\item{pathname}{Complete path (including filename) to your metabase
credentials}
}
\value{
mbcred - a list containing metabase credentials
}
\description{
This function reads credentials for an LTER Metabase into the mbcred
list variable.
}
