% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{migrate_eal_dir}
\alias{migrate_eal_dir}
\title{Migrate an EAL dataset directory to a jerald directory}
\usage{
migrate_eal_dir(eal.dir, jerald.dir)
}
\arguments{
\item{eal.dir}{Dataset directory, in EAL format, to migrate from}

\item{jerald.dir}{Dataset directory, in jerald format, to migrate to}
}
\description{
Create a new directory for a dataset with some template scripts,
metadata files, useful subdirectories, and a readme file. Template
files are in \verb{jerald/inst/template/}.
}
