% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{publish_dataset_edi}
\alias{publish_dataset_edi}
\title{Publish a dataset at the EDI repository}
\usage{
publish_dataset_edi(
  datasetid,
  mb.name,
  mb.cred,
  edi.cred,
  edi.env = "staging",
  dry.run = TRUE,
  s3.upload = TRUE,
  multi.part = FALSE,
  skip_checks = FALSE,
  bucket.name = Sys.getenv("AWS_S3_BUCKETNAME")
)
}
\arguments{
\item{datasetid}{ID number of the dataset to find in metabase and
update in EDI}

\item{mb.name}{name of the metabase database in the postgres cluster}

\item{mb.cred}{list of credentials for the metabase postgres cluster}

\item{edi.cred}{list of credentials to use for EDI}

\item{edi.env}{name of the EDI environment to update (staging, production,
or development)}

\item{dry.run}{boolean value - write EML only, then stop (end before s3
and EDI upload) if TRUE, continue to publish if FALSE}

\item{s3.upload}{boolean value (T/F) if TRUE upload to the s3 bucket, if
FALSE skip this (entities already there). Note that this does not currently
do a check on whether entities are present or not.}

\item{skip_checks}{boolean value (T/F) indicating whether or not to check
for congruence between data entity and attribute metadata
(check_attribute_congruence function). May want to set as True if the data
are online and not in the working directory.}

\item{bucket.name}{name of the s3 bucket to push data entities to}
}
\description{
This function publishes a dataset (or "data package") at the EDI
research data repository using metadata derived from an LTER Metabase.
The user must supply credentials for the metabase and EDI (see
\code{load_metabase_cred} and \code{load_destination_cred} functions), and
appropriate database and EDI environment names.
}
\details{
The basic process is to
\enumerate{
\item Pull metadata for the dataset from the metabase using \code{MetaEgress}
\item Query for the current dataset revision in the EDI environment
\item Write an EML document for the next revision to go to EDI
\item Push EML data entities from the working directory to an s3 bucket
\item Push the EML document to EDI, which triggers PASTA to pull data from
the s3 bucket and update the data package.
}
}
