% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{remove_ents_s3}
\alias{remove_ents_s3}
\title{Remove entities from S3 bucket}
\usage{
remove_ents_s3(ents, bucketname)
}
\arguments{
\item{ents}{A list of filenames}

\item{bucketname}{Name of the s3 bucket to upload to}
}
\description{
This function takes a list of data entities and uploads to an s3 bucket
}
