% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{template_dataset_dir}
\alias{template_dataset_dir}
\title{Create a dataset directory from a jerald template}
\usage{
template_dataset_dir(datasetid, get.edi = FALSE)
}
\arguments{
\item{datasetid}{ID number of the dataset to find in metabase and
update in EDI}
}
\description{
Create a new directory for a dataset with some template scripts,
metadata files, useful subdirectories, and a readme file. Template
files are in \verb{jerald/inst/template/}.
}
