% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{update_revnum_edi}
\alias{update_revnum_edi}
\title{Update EML revision number from EDI}
\usage{
update_revnum_edi(eml.list, edienv = "staging")
}
\arguments{
\item{eml.list}{EML metadata list}

\item{edienv}{Name of EDI repository environment ('staging', 'production',
or 'development')}
}
\value{
An EML-formatted list of metadata (with revised revnum)
}
\description{
This function queries the EDI repository to retrieve the current
revision number for a data package and updates an EML list accordingly
}
